feat: add application logger interface
This commit adds the `AppLogger` interface to the `logger.go` file in the `pkg/application` package. The interface defines methods for logging different levels of messages, such as `Info`, `Debug`, `Error`, and `Trace`. This interface will provide a standardized way for the application to log messages, allowing for easier testing and swapping of different logger implementations.